Fixed a bunch of bugs when creating new appcast files.
Bugs fixed include: the inability to edit appcast settings for a newly created file, inability to undo changes in a new file, not tracking whether a new document is dirty, undoing changes in a new document producing the text "<null>". This has introduced a few KVO errors (but no observable bugs) when opening an existing file.