Redesigned the appcast configuration sheet.
The settings are no longer kept on a sheet but are instead accessed via a disclosure triangle in the main window. This is better since it means the user can view the appcast configuration in the versions browser and can access undo/redo for the appcast configuration, something which was not possible using a modal sheet.